Stop subsequent execution in the case of errors in initramfs flash mode
The flash mode reboots the system after the new WIC image has been flashed successfully. The CI pipeline
detects the transition from reachable to unreachable DUT, in order to synchronise the subsequent processing.
In the case of errors, the initramfs flash mode stops the execution, which allows the CI pipeline to detect
this, because the DUT remains reachable. No reboot will be performed.