iot-hub-device-update: fix jq call in recipe
In appended installation (do_install:append(){}) call to jq with various
arguments to get passed on variables from the environent are placed as
parameters.

However, when such a variable isn't set the replacement yields the empty
string which is _NOT_ treated as the value for the parameter!
Therefore, place variables in double quotes to ensure that the --args
parameters are accompanied by a value.

Signed-off-by: Harry Waschkeit <harry.waschkeit@googlemail.com>