POC: add LTE support for phytec tauri
This is a first proof of concept to enable internet access for an omnect
device over LTE.

Most of the parameters necessary to generally bring up an LTE modem are
hard coded at the moment.

Besides, not a lot systematic tests were done conecrning start-up
stability, nor for long-term stability.

Quite some refinement will be required to make that feature
production-ready.