networkmanager: change configuration
Don't install a cellular configuration in system-connections anymore,
because it will depend on the application scenario and therefore needs
to be injected into the image as needed.

Additionally, a patch was added which re-enables IPv6 in case also IPv6
addresses get assigned during connection establishment.

And finally a tiny clean-up in NetworkManager.conf.

Signed-off-by: Harry Waschkeit <harry.waschkeit@googlemail.com>