u-boot: add patches to fix USB enumeration problem in rpi4
During startup in U-Boot USB interfaces are enumerated for boot devices.
However, if there are devices which aren't ready yet for operation these
caused the boot loader to panic and issue a reset, ending in a boot
loop.

Three patches found in different places fix that problem tin that way
that such devices get ignored.

Signed-off-by: Harry Waschkeit <harry.waschkeit@googlemail.com>