feat: introduce conventional commit messages
With this commit we introduce conventional commit messages (www.conventionalcommits.org).
In order to enforce the convention, a github action will check for a valid pull request title which serves as commit message title for the subsequent squash merge.
In accordance the pull request description is taken as commit message body of the squash merge.
A seperate CHANGELOG.md file is obsolet from now  on.