feat(ssh): let sshd determine ssh tunnel principal (#412)
We use the device id as principal for the ssh tunnel user. Until now we
therefore would inject a device id into the file. However, sshd can
itself retrieve this information via the AuthorizedPrincipalsCommand.
With this we adjust the configuration to use the
"omnect_get_deviceid.sh" script so that sshd can retrieve its config
itself.