feat: embed u-boot version in binary
Previously, bootloader version was stored separately from binary in
u-boot environment which is not really safe, because even if bootloader
itself wouldn't tamper with this setting - it is actually even
responsible for storing this information in an environment variable -
under Linux it is generally possible to delete or modify this variable.

To avoid this potantial problem and also the fact that a bootloader can
only update this information upon first run, the binary got extended to
bear the information.

With this apporach the version of the currently programmed bootloader
version can be retrieved under Linux as soon as it is flashed.

Signed-off-by: Harry Waschkeit <44188360+HarryWaschkeit@users.noreply.github.com>