feat(omnect-os-initramfs): in flash mode 2 allow direct flashing
Previously, the provided wic image was verified against the provided
bmap file which requires both of them stored in memory (tmpfs).

For devices with low memory this feature allows to directly program the
image into the flash device, of course accompanied by the danger to
brick a device in case a corrupted wic file is passed on to the device.

Default behavior is still to first verify, to enable direct flashing the
new bitbake variable OMNECT_FLASH_MODE_2_DIRECT_FLASHING needs to be set
to "1".

Signed-off-by: Harry Waschkeit <44188360+HarryWaschkeit@users.noreply.github.com>