refactor(initramfs): ensure unique partition UUIDs for flash-mode 1
Copied rootA and boot patrtitions had identical partition UUIDs on
source and destination medium.

This could lead to strange boot behavior, at least on x68 machines,
where either the wrong boot medium was used or booting was even
impossible after removal of the source medium.

To avoid that problem now generate a new partition UUID for boot as well
as rootA partitions and set them in the partition table accordingly.

Signed-off-by: Harry Waschkeit <44188360+HarryWaschkeit@users.noreply.github.com>